Explicitly add GSSCredential to InitialDirContext and close it explicitly
If the GSSCredential is obtained implicitly from the subject, referral contexts aren't able to authenticate with Kerberos because auth is not propagated. If the credential is added explicitly as SASL property, the environment is properly inherited and auth can succeed. When the context is closed by the caller, we explicitly dispose the GSSCredential to avoid any leakage.